// frontend/src/services/generated/models/ProgramDayResponse.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Fitlytics API
 * Workout tracking and training analytics API. Protected routes require a WorkOS access token in the `Authorization: Bearer ...` header.
 *
 * The version of the OpenAPI document: 0.1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { ProgramExerciseResponse } from './ProgramExerciseResponse';
import {
    ProgramExerciseResponseFromJSON,
    ProgramExerciseResponseFromJSONTyped,
    ProgramExerciseResponseToJSON,
    ProgramExerciseResponseToJSONTyped,
} from './ProgramExerciseResponse';

/**
 * 
 * @export
 * @interface ProgramDayResponse
 */
export interface ProgramDayResponse {
    /**
     * 
     * @type {Array<ProgramExerciseResponse>}
     * @memberof ProgramDayResponse
     */
    exercises?: Array<ProgramExerciseResponse>;
    /**
     * 
     * @type {string}
     * @memberof ProgramDayResponse
     */
    id?: string;
    /**
     * 
     * @type {boolean}
     * @memberof ProgramDayResponse
     */
    isRestDay?: boolean;
    /**
     * 
     * @type {string}
     * @memberof ProgramDayResponse
     */
    name?: string;
    /**
     * 
     * @type {string}
     * @memberof ProgramDayResponse
     */
    notes?: string;
    /**
     * 
     * @type {number}
     * @memberof ProgramDayResponse
     */
    sequence?: number;
    /**
     * 
     * @type {string}
     * @memberof ProgramDayResponse
     */
    tag?: string;
}

/**
 * Check if a given object implements the ProgramDayResponse interface.
 */
export function instanceOfProgramDayResponse(value: object): value is ProgramDayResponse {
    return true;
}

export function ProgramDayResponseFromJSON(json: any): ProgramDayResponse {
    return ProgramDayResponseFromJSONTyped(json, false);
}

export function ProgramDayResponseFromJSONTyped(json: any, ignoreDiscriminator: boolean): ProgramDayResponse {
    if (json == null) {
        return json;
    }
    return {
        
        'exercises': json['exercises'] == null ? undefined : ((json['exercises'] as Array<any>).map(ProgramExerciseResponseFromJSON)),
        'id': json['id'] == null ? undefined : json['id'],
        'isRestDay': json['is_rest_day'] == null ? undefined : json['is_rest_day'],
        'name': json['name'] == null ? undefined : json['name'],
        'notes': json['notes'] == null ? undefined : json['notes'],
        'sequence': json['sequence'] == null ? undefined : json['sequence'],
        'tag': json['tag'] == null ? undefined : json['tag'],
    };
}

export function ProgramDayResponseToJSON(json: any): ProgramDayResponse {
    return ProgramDayResponseToJSONTyped(json, false);
}

export function ProgramDayResponseToJSONTyped(value?: ProgramDayResponse | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'exercises': value['exercises'] == null ? undefined : ((value['exercises'] as Array<any>).map(ProgramExerciseResponseToJSON)),
        'id': value['id'],
        'is_rest_day': value['isRestDay'],
        'name': value['name'],
        'notes': value['notes'],
        'sequence': value['sequence'],
        'tag': value['tag'],
    };
}