// src/services/generated/models/ProgramSetTargetResponse.ts
export interface ProgramSetTargetResponse {
    /**
     * 
     * @type {number}
     * @memberof ProgramSetTargetResponse
     */
    capLoadKg?: number;
    /**
     * 
     * @type {string}
     * @memberof ProgramSetTargetResponse
     */
    id?: string;
    /**
     * 
     * @type {string}
     * @memberof ProgramSetTargetResponse
     */
    intensityText?: string;
    /**
     * 
     * @type {number}
     * @memberof ProgramSetTargetResponse
     */
    prescribedLoadKg?: number;
    /**
     * 
     * @type {string}
     * @memberof ProgramSetTargetResponse
     */
    prescribedLoadModifier?: string;
    /**
     * 
     * @type {number}
     * @memberof ProgramSetTargetResponse
     */
    prescribedRpe?: number;
    /**
     * 
     * @type {number}
     * @memberof ProgramSetTargetResponse
     */
    repsMax?: number;
    /**
     * 
     * @type {number}
     * @memberof ProgramSetTargetResponse
     */
    repsMin?: number;
    /**
     * 
     * @type {number}
     * @memberof ProgramSetTargetResponse
     */
    sequence?: number;
    /**
     * 
     * @type {string}
     * @memberof ProgramSetTargetResponse
     */
    setType?: string;
    /**
     * 
     * @type {number}
     * @memberof ProgramSetTargetResponse
     */
    setsCount?: number;
}